add: user block checker:
