+++ mini-app/src/cronJobs/index.ts
@@ -9,6 +9,7 @@ import { CreateRewards } from "@/cronJobs/tasks/CreateRewards";
 import { notifyUsersForRewards } from "@/cronJobs/tasks/notifyUsersForRewards";
 import { cronJobRunner } from "@/cronJobs/cornJobRunner";
 import { syncSbtCollectionsForEvents } from "@/cronJobs/tasks/syncSbtCollectionsForEvents";
+import { CheckAllUsersBlock } from "@/cronJobs/tasks/checkBlockStatus";
 
 const cronJobs = {
   CheckTransactions,
@@ -21,6 +22,7 @@ const cronJobs = {
   CreateRewards,
   notifyUsersForRewards,
   syncSbtCollectionsForEvents,
+  CheckAllUsersBlock,
 };
 export { cronJobRunner };
 export default cronJobs;
