+++ telegram-bot/src/controllers/sendMessage.ts
@@ -1,0 +1,99 @@
+import { Bot, GrammyError } from "grammy";
+import { logger } from "../utils/logger";
+import { Request, Response } from "express";
+
+interface SendRewardLinkBody {
+  chat_id: number;
+  link: string;
+  custom_message: string;
+}
+
+/**
+ * Sends a message to a Telegram chat.
+ *
+ * @param {Request & { bot: Bot }} req - The request object.
+ * @param {Response} res - The response object.
+ * @return {Promise<Response>} The response object.
+ */
+export const sendMessage = async (
+  req: Request & { bot: Bot },
+  res: Response,
+  tryCount?: number,
+): Promise<Response> => {
+  try {
+    // Destructure the request body
+    const { chat_id, link, custom_message }: SendRewardLinkBody = req.body;
+
+    // Input validation
+    if (!chat_id || typeof Number(chat_id) !== "number") {
+      // Return an error response if chat_id is invalid or missing
+      return res.status(400).json({ error: "Invalid or missing chat_id" });
+    }
+    if (link && (typeof link !== "string" || !link?.startsWith("http"))) {
+      // Return an error response if link is invalid
+      return res.status(400).json({ error: "Invalid link" });
+    }
+    if (!custom_message || typeof custom_message !== "string") {
+      // Return an error response if custom_message is missing or invalid
+      return res.status(400).json({ error: "Invalid custom_message" });
+    }
+
+    // Create the reply_markup object if link is provided
+    const reply_markup = link
+      ? {
+        inline_keyboard: [
+          [
+            {
+              text: "Claim Reward",
+              url: link,
+            },
+          ],
+        ],
+      }
+      : undefined;
+
+    // Send the message to the chat
+    await req.bot.api.sendMessage(chat_id, custom_message.trim(), {
+      reply_markup,
+      parse_mode: "HTML",
+    });
+
+    // Return a success response
+    return res.status(200).json({
+      success: true,
+      message: "Message sent successfully",
+    });
+  } catch (error) {
+    if (
+      error instanceof GrammyError &&
+      error.error_code === 429 &&
+      tryCount < 10
+    ) {
+      logger.error("BOT_ERROR: 429", error);
+
+      // wait 1000
+      await new Promise((resolve) => setTimeout(resolve, 1000));
+      await sendMessage(req, res, tryCount ? tryCount + 1 : 1);
+    }
+
+    // Differentiate between Telegram API errors and other errors
+    if (error.response?.statusCode) {
+      // Handle errors from the Telegram API (e.g., invalid chat_id, bot token issues)
+      const errorMessage =
+        error.response.description ||
+        "An error occurred while sending the message.";
+      logger.error("TELEGRAM_ERROR:", errorMessage);
+      return res.status(error.response.statusCode).json({
+        success: false,
+        error: errorMessage,
+      });
+    }
+
+    // Handle other unexpected errors
+    logger.error("Unexpected error sending message:", error);
+    return res.status(500).json({
+      success: false,
+      error: "Unexpected server error. Please try again later.",
+    });
+  }
+};
